사용하지 않는 command 삭제, TC 빌드 오류 수정, warning 제거

+++ SSD/VirtualSSD.cpp
@@ -217,7 +217,7 @@ void VirtualSSD::fetchDataFromNAND()
 {
 	ifstream file(NAND_FILE_NAME);
 	string line;
-	int commaIdx;
+	size_t commaIdx;
 
 	if (file.is_open()) {
 		while (!file.eof()) {
